Beginning modularization
Here is a very preliminary setup of some of the implementation that
needs to be done. I've created a few new functions, but there are many
more that need to be added. The Driver still contains some things that
it shouldn't, but at least it has some positive working dynamics for
determining conditions and actions that match the random market state.